requirements.txt
aiogram==3.3.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
SQLAlchemy>=2.0.0
asyncpg>=0.28.0
alembic>=1.12.0
databases>=0.8.0
aiosqlite>=0.19.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
celery[redis]>=5.3.0
pydantic>=2.4.1,<2.6
python-dateutil>=2.8.2
